requirements.txt
python-telegram-bot==20.0
openai==0.27.0
python-dotenv==0.19.2
pandas==1.3.5
matplotlib==3.5.1
pydub==0.25.1
SpeechRecognition==3.8.1